Create whitepaper on trauma and non-coding DNA
This whitepaper presents a scientific framework exploring the correlation between trauma, the Theory of Learning by Exception (TAE), and the activation of non-coding genomic regions, referred to as 'dark DNA'. It analyzes these concepts through the METFI model, proposing that trauma induces systemic reorganizations in human learning and biological coherence.